Merged PR 122805: Post Build Bot design and LUIS updates
Post Build Bot design and LUIS updates

Updates to Module 2.

Modified the "understanding the Organizational Requirement" slide and added content to the MD file
added a slide and md content on "Managing Bot States".
Added a customer story slide and md content for "Progressive"

Updates to module 3

Added a reference and MD content for LUDown
Added a reference and md content for chatdown
Added a slide and md content for the customer story Nedbank