Updates by CTO based on AnTho feedback
Hi, Chris. I read through the modules and I think for the most part the points are still valid. Just had a few items of potential feedback:

Bot design
Not sure if a Bot Directory still exists (Section 2.1). May need to recreate/alter this slide and the paragraph that follows

CTO: Updated the graphic to replace Bot Directory with Bot Integration

(Section 2.2) FormFlow doesn't exist in the bot builder SDK v4, need to look at the other tools available around prompting and dialog management

CTO: Replacing formflow with prompts
LUIS
Are we sure  "The intents and entities in a prebuilt domain are fully customizable once you've added them to your app - you can train them with utterances from your system, so they work for your users" this statement is true? I thought you couldn't update prebuilt entities or intents? (Section 1.1)

CTO:After rereading this, it is a mistake.

Changed it to

"The intents and entities in a custom domain are fully customizable once you've added them to your app - you can train them with utterances from your system, so they work for your users. You can use an entire prebuilt domain as a starting point for customization, or just borrow a few intents or entities from a prebuilt domain. "

Section 2.2: there are now many different types of entities, and I think perhaps we should include links to tutorials on how to use them and links on when to use them

CTO: I provided content on this in the PowerPoint and moved the content into the MD file. I don’t want to link of to a tutorial as it is a design course and I don't want to lose that design focus

Section 2.2: what this is referring to is unclear: There are many examples of the regular expressions in the /Specs directory for the supported cultures
Additionally, I think it'd be nice to include some examples of when you'd want to use exact match vs ML

CTO: Not sure what this is refering to

Section 3.1: I think we should share these tools "As a tool to help, you could use Chatdown with the V4 Emulator for designing sample conversations." but I am not sure where it should fit in this document
Here's a video about using chatdown to plan conversations https://channel9.msdn.com/Shows/AI-Show/Conversational-AI-Bot-Building-Tools

CTO: added the following to the MD and POwerrPOint file
As a tool to help, you could use [Chatdown](https://github.com/Microsoft/botbuilder-tools/tree/master/Chatdown) with the V4 [Emulator](https://docs.microsoft.com/en-us/azure/bot-service/bot-service-debug-emulator?view=azure-bot-service-4.0) for designing sample conversations. The following video about using chatdown to plan conversations can be found [here](https://channel9.msdn.com/Shows/AI-Show/Conversational-AI-Bot-Building-Tools).

Section 3.2: I think we should flesh this out a bit more, even if it's just including references to the documentation and related tutorials

CTO: Happy with how this is at the moment

Section 3.3: I think potentially we should remove recommending downloading the endpoint logs, I don't think it’s a good practice anymore

CTO: I still see this in the FAQ section
https://docs.microsoft.com/en-us/azure/cognitive-services/luis/luis-resources-faq#app-management

Section 4.1: Best Practices - potentially add some of the items I mention in the blog post - see the other page in this section of the notebook titled "anna's luis blog notes" - also here's the draft blog: LUIS: Notes from the Field

CTO: We can add this once it is live